Json data state has been moved to the HUB
